+++ tests/backends/test_capabilities.py
@@ -5,7 +5,8 @@ from tortoise.contrib import test
 class TestCapabilities(test.TestCase):
     # pylint: disable=E1101
 
-    def setUp(self):
+    async def asyncSetUp(self) -> None:
+        await super(TestCapabilities, self).asyncSetUp()
         self.db = Tortoise.get_connection("models")
         self.caps = self.db.capabilities