+++ tests/contrib/test_tester.py
@@ -22,10 +22,11 @@ class TestTesterSync(test.SimpleTestCase):
 
 
 class TestTesterASync(test.SimpleTestCase):
-    async def setUp(self):
+    async def asyncSetUp(self):
+        await super().asyncSetUp()
         self.baa = "TES"
 
-    async def tearDown(self):
+    def tearDown(self):
         self.assertEqual(self.baa, "TES")
 
     @test.skip("Skip it")