+++ tests/model_setup/test_init.py
@@ -6,7 +6,8 @@ from tortoise.exceptions import ConfigurationError
 
 
 class TestInitErrors(test.SimpleTestCase):
-    async def setUp(self):
+    async def asyncSetUp(self):
+        await super().asyncSetUp()
         try:
             Tortoise.apps = {}
             Tortoise._inited = False
@@ -14,8 +15,9 @@ class TestInitErrors(test.SimpleTestCase):
             pass
         Tortoise._inited = False
 
-    async def tearDown(self):
+    async def asyncTearDown(self) -> None:
         await Tortoise._reset_apps()
+        await super(TestInitErrors, self).asyncTearDown()
 
     async def test_basic_init(self):
         await Tortoise.init(