+++ tests/test_only.py
@@ -5,6 +5,7 @@ from tortoise.exceptions import IncompleteInstanceError
 
 class TestOnlyStraight(test.TestCase):
     async def asyncSetUp(self) -> None:
+        await super(TestOnlyStraight, self).asyncSetUp()
         self.model = StraightFields
         self.instance = await self.model.create(chars="Test")
 
@@ -60,5 +61,6 @@ class TestOnlyStraight(test.TestCase):
 
 class TestOnlySource(TestOnlyStraight):
     async def asyncSetUp(self) -> None:
+        await super().asyncSetUp()
         self.model = SourceFields  # type: ignore
         self.instance = await self.model.create(chars="Test")