+++ tortoise/expressions.py
@@ -1,3 +1,4 @@
+import operator
 from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Tuple, Type, Union, cast
 
 from pypika import Case as PypikaCase
@@ -248,7 +249,11 @@ class Q(Expression):
                     if param.get("value_encoder")
                     else model._meta.db.executor_class._field_to_db(field_object, value, model)
                 )
-            criterion = param["operator"](table[param["source_field"]], encoded_value)
+            op = param["operator"]
+            # this is an ugly hack
+            if op == operator.eq:
+                encoded_value = model._meta.db.query_class._builder()._wrapper_cls(encoded_value)
+            criterion = op(table[param["source_field"]], encoded_value)
         return criterion, join
 
     def _resolve_regular_kwarg(